+++ src/main/Extensions/WindowsControlPanel/WindowsControlPanel.test.ts
@@ -1,15 +1,17 @@
 import type { AssetPathResolver } from "@Core/AssetPathResolver/AssetPathResolver";
 import type { Translator } from "@Core/Translator";
+import type { SearchResultItem } from "@common/Core";
 import { describe, expect, it, vi } from "vitest";
 import { WindowsControlPanel } from "./WindowsControlPanel";
-import type { WindowsControlPanelItemsRepository } from "./WindowsControlPanelItemsRepositoryInterface";
+import type { WindowsControlPanelItem } from "./WindowsControlPanelItem";
+import type { WindowsControlPanelItemRepository } from "./WindowsControlPanelItemRepositoryInterface";
 
 describe(WindowsControlPanel, () => {
     describe(WindowsControlPanel.prototype.isSupported, () => {
         it("should only be supported on Windows", () => {
             const translator = <Translator>{};
             const assetPathResolver = <AssetPathResolver>{};
-            const repo = <WindowsControlPanelItemsRepository>{};
+            const repo = <WindowsControlPanelItemRepository>{};
 
             expect(new WindowsControlPanel("Windows", translator, assetPathResolver, repo).isSupported()).toBe(true);
             expect(new WindowsControlPanel("Linux", translator, assetPathResolver, repo).isSupported()).toBe(false);
@@ -23,39 +25,44 @@ describe(WindowsControlPanel, () => {
             const createTMock = vi.fn().mockReturnValue({ t });
             const translator = <Translator>{ createT: createTMock };
             const assetPathResolver = <AssetPathResolver>{};
-            const repo = <WindowsControlPanelItemsRepository>{
-                retrieveControlPanelItems: vi.fn().mockResolvedValue([
+
+            const windowsControlPanelItemRepository = <WindowsControlPanelItemRepository>{
+                retrieveControlPanelItems: vi.fn().mockResolvedValue(<WindowsControlPanelItem[]>[
                     {
                         Name: "item 1 name",
                         CanonicalName: "item 1 canonical name",
-                        Description: "item 1 description",
                         IconBase64: "item 1 icon",
                     },
                     {
                         Name: "item 2 name",
                         CanonicalName: "item 2 canonical name",
-                        Description: "item 2 description",
                         IconBase64: "item 2 icon",
                     },
                 ]),
             };
-            const extension = new WindowsControlPanel("Windows", translator, assetPathResolver, repo);
+
+            const extension = new WindowsControlPanel(
+                "Windows",
+                translator,
+                assetPathResolver,
+                windowsControlPanelItemRepository,
+            );
+
             extension["knownControlPanelItems"] = [
                 {
                     Name: "item 1 name",
                     CanonicalName: "item 1 canonical name",
-                    Description: "item 1 description",
                     IconBase64: "item 1 icon",
                 },
             ];
 
             const items = await extension.getSearchResultItems();
 
-            expect(items).toEqual([
+            expect(items).toEqual(<SearchResultItem[]>[
                 {
                     id: "item 1 canonical name",
                     name: "item 1 name",
-                    description: "item 1 description",
+                    description: t("searchResultItemDescription"),
                     image: { url: `data:image/png;base64,item 1 icon` },
                     defaultAction: {
                         argument: "item 1 name",
@@ -67,7 +74,7 @@ describe(WindowsControlPanel, () => {
                 {
                     id: "item 2 canonical name",
                     name: "item 2 name",
-                    description: "item 2 description",
+                    description: t("searchResultItemDescription"),
                     image: { url: `data:image/png;base64,item 2 icon` },
                     defaultAction: {
                         argument: "item 2 name",
@@ -77,11 +84,13 @@ describe(WindowsControlPanel, () => {
                     },
                 },
             ]);
-            expect(repo.retrieveControlPanelItems).toHaveBeenCalledWith([
+
+            expect(windowsControlPanelItemRepository.retrieveControlPanelItems).toHaveBeenCalledWith(<
+                WindowsControlPanelItem[]
+            >[
                 {
                     Name: "item 1 name",
                     CanonicalName: "item 1 canonical name",
-                    Description: "item 1 description",
                     IconBase64: "item 1 icon",
                 },
             ]);
