+++ src/main/Extensions/WindowsControlPanel/WindowsControlPanelItem.ts
@@ -1,6 +1,5 @@
 export type WindowsControlPanelItem = {
     Name: string;
     CanonicalName: string;
-    Description: string;
     IconBase64: string;
 };
