+++ src/main/Extensions/WindowsControlPanel/WindowsControlPanelItemRepositoryInterface.ts
@@ -1,0 +1,5 @@
+import type { WindowsControlPanelItem } from "./WindowsControlPanelItem";
+
+export interface WindowsControlPanelItemRepository {
+    retrieveControlPanelItems(alreadyKnownItems: WindowsControlPanelItem[]): Promise<WindowsControlPanelItem[]>;
+}
