--- src/main/Extensions/WindowsControlPanel/WindowsControlPanelItemsRepositoryInterface.ts
@@ -1,5 +1,0 @@
-import { WindowsControlPanelItem } from "./WindowsControlPanelItem";
-
-export interface WindowsControlPanelItemsRepository {
-    retrieveControlPanelItems(alreadyKnownItems: WindowsControlPanelItem[]): Promise<WindowsControlPanelItem[]>;
-}
